requirements.txt
asgiref==3.3.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorlog==4.6.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==4.4.2
Django==3.1.3
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
filetype==1.0.7
har2case==0.3.1
httprunner==2.3.0
idna==2.10
importlib-metadata==3.3.0
inflection==0.5.1
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonpath==0.82
Markdown==3.3.3
MarkupSafe==1.1.1
mysqlclient==2.0.1
packaging==20.8
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
